refactor: split daejeon event sync upload and reporting
